+++ src/test/java/com/motaharinia/msutility/custom/customjson/JsonUnitTest.java
@@ -22,7 +22,6 @@ import static org.assertj.core.api.Assertions.fail;
 
 /**
  * @author https://github.com/motaharinia<br>
- * Description: <br>
  * کلاس تست SearchDataDto
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)